Deboge + securation AddEquipes(Need login now for add a equipe)

+++ src/main/java/be/heh/projettrica/domain/entities/Equipes.java
@@ -22,5 +22,5 @@ public class Equipes {
     private final int nbwin;
 
     @Getter
-    private final int idequipe;
+    private final long idequipe;
 }
